+++ indico/modules/events/abstracts/controllers/abstract_list.py
@@ -10,8 +10,10 @@ from operator import attrgetter
 
 from flask import flash, jsonify, request, session
 from sqlalchemy.orm import joinedload, subqueryload
+from webargs import fields
 
 from indico.core.db import db
+from indico.core.db.sqlalchemy.util.models import get_simple_column_attrs
 from indico.modules.events.abstracts.controllers.base import RHManageAbstractsBase
 from indico.modules.events.abstracts.controllers.common import (AbstractsDownloadAttachmentsMixin, AbstractsExportCSV,
                                                                 AbstractsExportExcel, AbstractsExportPDFMixin,
@@ -28,6 +30,7 @@ from indico.modules.events.contributions.models.persons import AuthorType
 from indico.modules.events.util import get_field_values
 from indico.modules.users.models.users import User
 from indico.util.i18n import _, ngettext
+from indico.web.args import use_kwargs
 from indico.web.util import jsonify_data, jsonify_form, jsonify_template
 
 
@@ -107,11 +110,37 @@ class RHAbstractListStaticURL(RHAbstractListBase):
 
 
 class RHCreateAbstract(RHAbstractListBase):
+    """Create an abstract from scratch or with existing cloned data."""
+
+    def clone_fields(self, abstract):
+        field_names = ['title', 'description', 'submission_comment', 'submitted_for_tracks', 'submitted_contrib_type']
+        field_data = {f: getattr(abstract, f) for f in field_names}
+        person_links = []
+        link_attrs = get_simple_column_attrs(AbstractPersonLink)
+        for old_link in abstract.person_links:
+            link = AbstractPersonLink(person=old_link.person)
+            link.populate_from_attrs(old_link, link_attrs)
+            person_links.append(link)
+        field_data['person_links'] = person_links
+        for f in abstract.field_values:
+            field_data[f'custom_{f.contribution_field_id}'] = f.data
+        return field_data
+
+    @use_kwargs({
+        'abstract_id': fields.Int(load_default=None)
+    }, location='query')
+    def _process_args(self, abstract_id):
+        RHAbstractListBase._process_args(self)
+        self.abstract = None
+        if abstract_id:
+            self.abstract = Abstract.query.with_parent(self.event).filter_by(id=abstract_id).first_or_404()
+
     def _process(self):
         is_invited = request.args.get('invited') == '1'
         abstract_form_class = make_abstract_form(self.event, session.user, notification_option=True,
                                                  management=self.management, invited=is_invited)
-        form = abstract_form_class(event=self.event, management=self.management, invited=is_invited)
+        cloned_fields = self.clone_fields(self.abstract) if self.abstract else {}
+        form = abstract_form_class(event=self.event, management=self.management, invited=is_invited, **cloned_fields)
         if is_invited:
             del form.submitted_contrib_type
             del form.attachments
@@ -144,7 +173,8 @@ class RHCreateAbstract(RHAbstractListBase):
             if tpl_components.get('hide_abstract'):
                 self.list_generator.flash_info_message(abstract)
             return jsonify_data(**tpl_components)
-        return jsonify_form(form, back=_('Cancel'), form_header_kwargs={'action': request.relative_url})
+        return jsonify_form(form, back=_('Cancel'), disabled_until_change=(not self.abstract or is_invited),
+                            form_header_kwargs={'action': request.relative_url})
 
 
 class RHDeleteAbstracts(RHManageAbstractsActionsBase):